Update tag_deck.py to edit mod
Add lines to update `mod` property of cards to current time in epoch seconds. (In contrast, the version of the deck living in Anki would have the time set to when the deck was last edited).

This permits Anki to recognize that cards have new edits (because the mod time is later than what exists in Anki) and recognize that it needs to update tags.